Import Calendars from iCal file (.ics file) #114
* moved the event processing logic from ImportJob class to the helper CalendarEventImporter class
