+++ application-mocca-calendar-api/src/main/java/org/xwiki/contrib/moccacalendar/internal/importJob/ImportJob.java
@@ -19,40 +19,27 @@
  */
 package org.xwiki.contrib.moccacalendar.internal.importJob;
 
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Provider;
 
-import org.jsoup.Jsoup;
-import org.jsoup.safety.Safelist;
 import org.xwiki.component.annotation.Component;
 import org.xwiki.contrib.moccacalendar.importJob.ImportJobRequest;
 import org.xwiki.contrib.moccacalendar.importJob.ImportJobStatus;
-import org.xwiki.contrib.moccacalendar.internal.EventConstants;
 import org.xwiki.fullcalendar.FullCalendarManager;
-import org.xwiki.fullcalendar.model.MoccaCalendarEvent;
-import org.xwiki.fullcalendar.model.RecurrentEventModification;
+import org.xwiki.fullcalendar.model.CalendarEvent;
 import org.xwiki.job.AbstractJob;
 import org.xwiki.job.GroupedJob;
 import org.xwiki.job.JobGroupPath;
-import org.xwiki.model.reference.DocumentReference;
-import org.xwiki.model.reference.DocumentReferenceResolver;
-import org.xwiki.model.reference.SpaceReference;
-import org.xwiki.rendering.syntax.Syntax;
 import org.xwiki.stability.Unstable;
-import org.xwiki.wysiwyg.converter.HTMLConverter;
 
 import com.xpn.xwiki.XWiki;
 import com.xpn.xwiki.XWikiContext;
 import com.xpn.xwiki.XWikiException;
 import com.xpn.xwiki.doc.XWikiDocument;
-import com.xpn.xwiki.objects.BaseObject;
 
 /**
  * The Mocca calendar import job.
@@ -70,12 +57,6 @@ public class ImportJob extends AbstractJob<ImportJobRequest, ImportJobStatus> im
      */
     public static final String JOB_TYPE = "moccacalendar.import";
 
-    private static final String NEW_EVENT_HOME = "WebHome";
-
-    @Inject
-    @Named("current")
-    private DocumentReferenceResolver<String> documentReferenceResolver;
-
     @Inject
     private Provider<XWikiContext> wikiContextProvider;
 
@@ -83,7 +64,7 @@ public class ImportJob extends AbstractJob<ImportJobRequest, ImportJobStatus> im
     private FullCalendarManager fullCalendarManager;
 
     @Inject
-    private HTMLConverter htmlConverter;
+    private CalendarEventImporter calendarEventImporter;
 
     @Override
     public String getType()
@@ -111,12 +92,12 @@ public class ImportJob extends AbstractJob<ImportJobRequest, ImportJobStatus> im
     protected void runInternal()
     {
         try {
-            List<MoccaCalendarEvent> calendarEventsJson =
+            List<CalendarEvent> calendarEventsJson =
                 fullCalendarManager.getICalEventsFromFile(getICSFileContent(request.getFile()), null, null, true);
             List<XWikiDocument> eventDocuments = new ArrayList<>();
             this.progressManager.pushLevelProgress(calendarEventsJson.size() + 1, this);
 
-            for (MoccaCalendarEvent calendarEvent : calendarEventsJson) {
+            for (CalendarEvent calendarEvent : calendarEventsJson) {
                 progressManager.startStep(this);
                 if (status.isCanceled()) {
                     progressManager.endStep(this);
@@ -124,8 +105,9 @@ public class ImportJob extends AbstractJob<ImportJobRequest, ImportJobStatus> im
                 }
 
                 XWikiDocument eventDoc =
-                    getUniqueEventName(calendarEvent.getTitle().trim(), request.getParentRef(), eventDocuments);
-                createCalendarObjects(eventDoc, calendarEvent);
+                    calendarEventImporter.getUniqueEventName(calendarEvent.getTitle().trim(), request.getParentRef(),
+                        eventDocuments);
+                calendarEventImporter.createCalendarObjects(eventDoc, calendarEvent);
                 eventDocuments.add(eventDoc);
 
                 progressManager.endStep(this);
@@ -152,75 +134,6 @@ public class ImportJob extends AbstractJob<ImportJobRequest, ImportJobStatus> im
         }
     }
 
-    private void createCalendarObjects(XWikiDocument eventDoc, MoccaCalendarEvent component) throws XWikiException
-    {
-        XWikiContext wikiContext = wikiContextProvider.get();
-        DocumentReference eventClassRef =
-            documentReferenceResolver.resolve(EventConstants.MOCCA_CALENDAR_EVENT_CLASS_NAME);
-        BaseObject eventObj = eventDoc.newXObject(eventClassRef, wikiContext);
-
-        addConvertedPropertyToObject(component.getDescription(), eventObj, Syntax.XWIKI_2_1.toIdString(),
-            EventConstants.PROPERTY_DESCRIPTION_NAME);
-        addConvertedPropertyToObject(component.getTitle(), eventObj, Syntax.XWIKI_2_1.toIdString(),
-            EventConstants.PROPERTY_TITLE_NAME);
-
-        int allDay = component.isAllDay() ? 1 : 0;
-        eventObj.set(EventConstants.PROPERTY_ALLDAY_NAME, allDay, wikiContext);
-        if (allDay == 1) {
-            LocalDateTime end = component.getEnd().toInstant().atZone(ZoneOffset.UTC).toLocalDateTime().minusDays(1);
-            eventObj.set(EventConstants.PROPERTY_ENDDATE_NAME, Date.from(end.atZone(ZoneOffset.UTC).toInstant()),
-                wikiContext);
-        } else {
-            eventObj.set(EventConstants.PROPERTY_ENDDATE_NAME, component.getEnd(), wikiContext);
-        }
-        eventObj.set(EventConstants.PROPERTY_STARTDATE_NAME, component.getStart(), wikiContext);
-
-        int recurrenceValue = component.isRecurrent();
-        eventObj.set(EventConstants.PROPERTY_RECURRENT_NAME, recurrenceValue, wikiContext);
-        if (recurrenceValue == 1) {
-            processRecurrence(eventDoc, component);
-        }
-    }
-
-    private void addConvertedPropertyToObject(String content, BaseObject eventObj, String syntax, String property)
-    {
-        String cleanContent = Jsoup.clean(content, Safelist.basic());
-        String convertedContent = htmlConverter.fromHTML(cleanContent, syntax);
-
-        eventObj.set(property, convertedContent, wikiContextProvider.get());
-    }
-
-    private void processRecurrence(XWikiDocument eventDoc, MoccaCalendarEvent component) throws XWikiException
-    {
-        XWikiContext wikiContext = wikiContextProvider.get();
-        DocumentReference eventRecClassRef =
-            documentReferenceResolver.resolve(EventConstants.MOCCA_CALENDAR_EVENT_RECURRENCY_CLASS_NAME);
-        BaseObject eventRecObj = eventDoc.newXObject(eventRecClassRef, wikiContext);
-
-        eventRecObj.set(EventConstants.PROPERTY_FIRSTINSTANCE_NAME, component.getStart(), wikiContext);
-        eventRecObj.set(EventConstants.PROPERTY_LASTINSTANCE_NAME, component.getRecEndDate(), wikiContext);
-        eventRecObj.set(EventConstants.PROPERTY_FREQUENCY_NAME, component.getRecurrenceFreq().toLowerCase(),
-            wikiContext);
-
-        List<RecurrentEventModification> modList = component.getModificationList();
-        DocumentReference eventRecModifiedRef =
-            documentReferenceResolver.resolve(EventConstants.MOCCA_CALENDAR_EVENT_MODIFICATION_CLASS_NAME);
-
-        for (RecurrentEventModification eventModification : modList) {
-            BaseObject eventModObj = eventDoc.newXObject(eventRecModifiedRef, wikiContext);
-            eventModObj.set(EventConstants.PROPERTY_ORIG_STARTDATE_OF_MODIFIED_NAME,
-                eventModification.getOriginalDate(), wikiContext);
-            eventModObj.set(EventConstants.PROPERTY_STARTDATE_NAME, eventModification.getModifiedStartDate(),
-                wikiContext);
-            eventModObj.set(EventConstants.PROPERTY_ENDDATE_NAME, eventModification.getModifiedEndDate(), wikiContext);
-            addConvertedPropertyToObject(eventModification.getModifiedTitle(), eventModObj,
-                Syntax.XWIKI_2_1.toIdString(), EventConstants.PROPERTY_TITLE_NAME);
-
-            addConvertedPropertyToObject(eventModification.getModifiedDescription(), eventModObj,
-                Syntax.XWIKI_2_1.toIdString(), EventConstants.PROPERTY_DESCRIPTION_NAME);
-        }
-    }
-
     private byte[] getICSFileContent(byte[] importedFileContent)
     {
         String importedFileString = new String(importedFileContent);
@@ -230,37 +143,5 @@ public class ImportJob extends AbstractJob<ImportJobRequest, ImportJobStatus> im
         String filteredCalendarContent = importedFileString.substring(beginIndex, endIndex);
 
         return filteredCalendarContent.getBytes();
-    }
-
-    private XWikiDocument getUniqueEventName(String eventName, String parentRef, List<XWikiDocument> eventDocuments)
-        throws XWikiException
-    {
-        XWikiContext wikiContext = this.wikiContextProvider.get();
-
-        DocumentReference parentDocRef = documentReferenceResolver.resolve(parentRef);
-        SpaceReference documentSpaceRef = new SpaceReference(eventName, parentDocRef.getLastSpaceReference());
-        DocumentReference eventRef = new DocumentReference(NEW_EVENT_HOME, documentSpaceRef);
-        wikiContext.getWiki().getDocument(eventRef, wikiContext);
-        XWikiDocument eventDoc = wikiContext.getWiki().getDocument(eventRef, wikiContext);
-
-        if (!eventDoc.isNew() || sameNameEventExists(eventRef, eventDocuments)) {
-            String newEventName = eventName + "_" + System.nanoTime();
-            SpaceReference newDocumentSpaceRef = new SpaceReference(newEventName, parentDocRef.getLastSpaceReference());
-            DocumentReference newEventRef = new DocumentReference(NEW_EVENT_HOME, newDocumentSpaceRef);
-            eventDoc = new XWikiDocument(newEventRef);
-        }
-        eventDoc.setTitle(eventName);
-        eventDoc.setParentReference(parentDocRef);
-        return eventDoc;
-    }
-
-    private boolean sameNameEventExists(DocumentReference eventRef, List<XWikiDocument> eventDocuments)
-    {
-        for (XWikiDocument document : eventDocuments) {
-            if (document.getDocumentReference().equals(eventRef)) {
-                return true;
-            }
-        }
-        return false;
     }
 }
