+++ application-mocca-calendar-api/src/main/resources/META-INF/components.txt
@@ -13,5 +13,6 @@ org.xwiki.contrib.moccacalendar.internal.generators.YearlyEventGenerator
 org.xwiki.contrib.moccacalendar.internal.meetings.MeetingEventSource
 org.xwiki.contrib.moccacalendar.internal.meetings.MeetingsSourceConfigurationClassInitializer
 org.xwiki.contrib.moccacalendar.internal.utils.DefaultEventAssembly
+org.xwiki.contrib.moccacalendar.internal.importJob.CalendarEventImporter
 org.xwiki.contrib.moccacalendar.internal.importJob.ImportJob
 org.xwiki.contrib.moccacalendar.internal.rest.DefaultMoccaCalendarResource
